+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from 'react';
-import { Box, Spinner, Stack } from '@chakra-ui/react';
+import { Box, Spinner, Stack, useDisclosure } from '@chakra-ui/react';
 
 import Header from './components/layout/Header';
 import Main from './components/layout/Main';
@@ -7,11 +7,14 @@ import Footer from './components/layout/Footer';
 
 import { SearchContextProvider } from './context/SearchContext';
 import useViewer from './hooks/queries/useViewer';
+import SearchDrawerContext from './context/SearchDrawerContext';
 
 export default function App() {
 	const {
 		result: { loading },
 	} = useViewer();
+
+	const { isOpen: drawerIsOpen, onOpen: openDrawer, onClose: closeDrawer } = useDisclosure();
 
 	// Get the header height so we can offset the main content
 	const headerRef = useRef(null);
@@ -49,11 +52,13 @@ export default function App() {
 					color: 'text.dark',
 				}}
 			>
-				<Header ref={headerRef} />
-				<Box minH='66vh' w='full' paddingTop={`${headerHeight}px`}>
-					{loading ? <Spinner /> : <Main />}
-				</Box>
-				<Footer />
+				<SearchDrawerContext.Provider value={{ drawerIsOpen, openDrawer, closeDrawer }}>
+					<Header ref={headerRef} />
+					<Box minH='66vh' w='full' paddingTop={`${headerHeight}px`}>
+						{loading ? <Spinner /> : <Main />}
+					</Box>
+					<Footer />
+				</SearchDrawerContext.Provider>
 			</Stack>
 		</SearchContextProvider>
 	);