+++ src/views/ProfileView.tsx
@@ -37,17 +37,17 @@ import ReactPlayer from 'react-player';
 import { getWPItemsFromIds } from '../lib/utils';
 import { Credit, UserProfile, WPItem } from '../lib/classes';
 import { useProfileUrl } from '../hooks/hooks';
+import useViewer from '../hooks/queries/useViewer';
 import useUserTaxonomies from '../hooks/queries/useUserTaxonomies';
 import BookmarkToggleIcon from '../components/common/BookmarkToggleIcon';
+import LinkWithIcon from '../components/common/LinkWithIcon';
+import ShareButton from '../components/common/ShareButton';
+import WrapWithIcon from '../components/common/WrapWithIcon';
+import ProfileDisabledNotice from '../components/common/ProfileDisabledNotice';
 import CreditsTagLegend from '../components/CreditsTagLegend';
 import PersonalIconLinks from '../components/PersonalIconLinks';
 import CreditItem from '../components/CreditItem';
 import HeadingCenterline from '../components/common/HeadingCenterline';
-import LinkWithIcon from '../components/common/LinkWithIcon';
-import ShareButton from '../components/common/ShareButton';
-import WrapWithIcon from '../components/common/WrapWithIcon';
-import useViewer from '../hooks/queries/useViewer';
-import ProfileDisabledNotice from '../components/common/ProfileDisabledNotice';
 
 interface Props {
 	profile: UserProfile;